fix job_composer and model

+++ app/model.py
@@ -1,6 +1,8 @@
+from pathlib import Path
 from pydantic import BaseModel, Field
 from typing import List, Optional
 from datetime import datetime
+
 
 class Resolution(BaseModel):
     width_px: int = Field(..., description="Width in pixels")
@@ -32,15 +34,16 @@ class Environment(BaseModel):
     cpu_name: str
     cpu_threads: int
 
+
 class SourceVideo(BaseModel):
     file_name: str
     file_size_megabytes: float
     resolution: Resolution
     video_duration_seconds: float
     codec: str
-    average_bitrate_kilobits_per_second: int
+    average_bitrate_kilobits_per_second: float
     fps: float
-    actual_frame_count: int # Can be calculated: fps * duration
+    actual_frame_count: int  # Can be calculated: fps * duration
     sha256_hash: str
     ffmpeg_metadata: FfmpegMetadata
 
@@ -59,7 +62,7 @@ class Iteration(BaseModel):
     file_size_megabytes: float
     video_duration_seconds: float
     codec: str
-    encoder_settings: EncoderSettings  
+    encoder_settings: EncoderSettings
     source_to_encoded_vmaf_percent: float
 
     encoding_finished_datetime: datetime
@@ -68,13 +71,15 @@ class Iteration(BaseModel):
     actual_frame_count: int
     sha256_hash: str
     ffmpeg_command_used: str
-    environment: Environment  
+    environment: Environment
     ffmpeg_metadata: FfmpegMetadata
+
 
 class EncoderDataJson(BaseModel):
     source_video: SourceVideo
     encoding_stage: EncodingStage
     iterations: List[Iteration]
+
 
 class EncoderJobContext(BaseModel):
     source_file_path: Path
